+++ src/components/myPage/tripCard.tsx
